.fp: Allow spaces in filenames
The documentation stated that %20 could be used to escape spaces.
However this was not implemented.  The documentation and implementation
has now been changed to allow spaces by quoting the filename with '"'.

This can also be used for .cf, .lf, .nx, and .so now.

Additionally character codes up to 255 are allowed now in filenames.