.tr bugfix: Ignore char 2 if char 1 is unknown
This is an old roff bug:  If the first character in the .tr argument
is unknown it is skipped and the second character becomes the first.
Now all translations are done wrong.
This bug does only occur if the first character is unknown.  An
unknown character at a other position is treated correctly.