Fix listing fonts for -ms
The L font family has not existed for a while now.
It used to be Geneva Light, a font similar to Helvetica.
Because font licensing is difficult, getting hold of a real Geneva Light
is most likely not an actual option.

It's somewhat questionable whether sans-serif fonts are a good choice
for listings, but historical accuracy is more important.

This fixes #66.