pic: "chop" does not work
In commit c6c12d9... dated 04 Nov 2014, "Some gcc warnings removed,"
the line
chop_ap[chop++] = ap;
was deleted from file pic/linegen.c.
gcc -Wall does flag this line with the message that chop_ap is
an unused variable. It is right about that. However, the variable
chop is used, and if it is not incremented, it remains at 0 and
the "chop" attribute won't work.
Solution: Add chop++ back in.