Fix incorrectly classified languages on GitHub
This commit corrects the categorisation and highlighting of unrecognised
filetypes on GitHub.com. The library used on GitHub for identifying file
formats (https://github.com/github/linguist) relies heavily on filenames
to identify filetypes. As this approach is far from foolproof, overrides
are provided to authors in the form of a gitattributes(5) directive:

    *.foo linguist-language=C    # Specify language
    *.bar linguist-vendored      # Exclude from language statistics

Only supported formats (those listed in Linguist's `languages.yml` file)
can be declared this way. A common workaround for misclassified files is
to flag them as "Text", which is effectively the logical default anyway.

See: https://github.com/github/linguist/blob/cfafb37b/docs/overrides.md