Synchronise editor modelines with `.gitattributes`
This ensures that the appearance of files stays consistent between one's
editor and how it's displayed on GitHub.com. Plain-text overrides aren't
as relevant to text-editors, so we only sync `linguist-language=Text` if
the corresponding file would be interpreted by editors as something that
isn't ordinary text.

Note that some of modelines (mostly Vim's) are technically redundant; to
enforce consistency and obviate cross-checking between Vim and Emacs, we
define both `-*- name -*-` and `vim: ft=name` together. Moreover, moving
them to the top of the file (instead of the bottom) makes them easier to
update via scripts, and reduces the likelihood of a Vim modeline falling
outside the scope configured locally by `:set modelines`.